requirements.txt
Flask
python-dotenv
Flask-Bcrypt
Flask-JWT-Extended
Flask-Limiter
pytest